Add socket module
