+++ src/epmd.rs
@@ -5,90 +5,13 @@
 //!
 //! See [EPMD Protocol](https://www.erlang.org/doc/apps/erts/erl_dist_protocol.html#epmd-protocol)
 //! for more details about the protocol.
+use crate::socket::Socket;
 use crate::Creation;
-use byteorder::{BigEndian, ByteOrder as _};
-use futures::io::{AsyncRead, AsyncReadExt as _, AsyncWrite, AsyncWriteExt as _};
+use futures::io::{AsyncRead, AsyncWrite};
 use std::str::FromStr;
 
 /// The default listening port of the EPMD.
 pub const DEFAULT_EPMD_PORT: u16 = 4369;
-
-// TODO: move module
-#[derive(Debug)]
-pub struct Socket<T> {
-    inner: T,
-}
-
-impl<T> Socket<T>
-where
-    T: AsyncRead + AsyncWrite + Unpin,
-{
-    pub fn new(inner: T) -> Self {
-        Self { inner }
-    }
-
-    pub fn into_inner(self) -> T {
-        self.inner
-    }
-
-    async fn write_u8(&mut self, v: u8) -> std::io::Result<()> {
-        self.inner.write_all(&[v]).await
-    }
-
-    async fn write_u16(&mut self, v: u16) -> std::io::Result<()> {
-        let mut buf = [0; 2];
-        BigEndian::write_u16(&mut buf, v);
-        self.inner.write_all(&buf).await
-    }
-
-    async fn write_all(&mut self, buf: &[u8]) -> std::io::Result<()> {
-        self.inner.write_all(&buf).await
-    }
-
-    async fn flush(&mut self) -> std::io::Result<()> {
-        self.inner.flush().await
-    }
-
-    async fn read_u8(&mut self) -> std::io::Result<u8> {
-        let mut buf = [0; 1];
-        self.inner.read_exact(&mut buf).await?;
-        Ok(buf[0])
-    }
-
-    async fn read_u16(&mut self) -> std::io::Result<u16> {
-        let mut buf = [0; 2];
-        self.inner.read_exact(&mut buf).await?;
-        Ok(BigEndian::read_u16(&buf))
-    }
-
-    async fn read_u32(&mut self) -> std::io::Result<u32> {
-        let mut buf = [0; 4];
-        self.inner.read_exact(&mut buf).await?;
-        Ok(BigEndian::read_u32(&buf))
-    }
-
-    async fn read_string(&mut self) -> std::io::Result<String> {
-        let mut buf = String::new();
-        self.inner.read_to_string(&mut buf).await?;
-        Ok(buf)
-    }
-
-    async fn read_u16_bytes(&mut self) -> std::io::Result<Vec<u8>> {
-        let mut buf = vec![0; usize::from(self.read_u16().await?)];
-        self.inner.read_exact(&mut buf).await?;
-        Ok(buf)
-    }
-
-    async fn read_u16_string(&mut self) -> std::io::Result<String> {
-        let buf = self.read_u16_bytes().await?;
-        String::from_utf8(buf).map_err(|_| {
-            std::io::Error::new(
-                std::io::ErrorKind::InvalidData,
-                "stream did not contain valid UTF-8",
-            )
-        })
-    }
-}
 
 const TAG_DUMP_REQ: u8 = 100;
 const TAG_KILL_REQ: u8 = 107;
@@ -293,20 +216,6 @@ where
             tag => Err(EpmdError::UnknownResponseTag { tag }),
         }
     }
-    //             .and_then(|(stream, _)| {
-    //                 let to_creation = |c| {
-    //                     Creation::from_u16(c).ok_or_else(|| invalid_data!("Too large creation: {}", c))
-    //                 };
-    //                 (
-    //                     U8.expect_eq(TAG_ALIVE2_RESP),
-    //                     U8.expect_eq(0),
-    //                     U16.be().and_then(to_creation),
-    //                 )
-    //                     .read_from(stream)
-    //             })
-    //             .map(|(stream, (_, _, creation))| (stream, creation))
-    //             .map_err(|e| e.into_error())
-    //     }
 
     /// Gets all registered names from EPMD.
     pub async fn get_names(mut self) -> Result<Vec<NodeName>, EpmdError> {
